+++ go-app/frontend/src/tabs/cabdebugger/CabDebuggerTab.tsx
@@ -4,6 +4,8 @@ import { useCabControlState } from "../../swr";
 import { CabDebuggerTabControl } from "./CabDebuggerTabControl";
 import { main } from "../../../wailsjs/go/models";
 import { CabDebuggerMapControlModal } from "./MapControlModal/CabDebuggerMapControlModal";
+import { CabDebuggerMapKeybindingModal } from "./MapKeybindingModal/CabDebuggerMapKeybindingModal";
+import { TiCog } from "react-icons/ti";
 
 export const CabDebuggerTab = () => {
   const { register, watch } = useForm<{ query: string }>({
@@ -13,6 +15,7 @@ export const CabDebuggerTab = () => {
     useCabControlState();
   const [mapControlModalOpenState, setMapControlModalOpenState] =
     useState<main.Interop_Cab_ControlState_Control | null>(null);
+  const [mapKeybindingModalOpen, setMapKeybindingModalOpen] = useState(false);
 
   const query = watch("query");
   const sortedControls = useMemo(
@@ -40,6 +43,10 @@ export const CabDebuggerTab = () => {
     setMapControlModalOpenState(null);
   }, []);
 
+  const handleCloseMapKeybindingModal = useCallback(() => {
+    setMapKeybindingModalOpen(false);
+  }, []);
+
   useEffect(() => {
     let interval: ReturnType<typeof setInterval> | null = null;
     interval = setInterval(() => {
@@ -57,6 +64,31 @@ export const CabDebuggerTab = () => {
           <div>Currently driving {cabControlState.Name}</div>
         </div>
       )}
+      <div className="grid items-center grid-cols-[minmax(0,1fr)_max-content] gap-2">
+        <input
+          className="input w-full"
+          placeholder="Search for control(s)"
+          {...register("query")}
+        />
+        <div>
+          <div className="dropdown dropdown-bottom dropdown-end">
+            <div tabIndex={0} role="button" className="btn py-4 px-2">
+              <TiCog size={20} />
+            </div>
+            <ul
+              tabIndex={-1}
+              className="dropdown-content menu bg-base-300 rounded-box z-1 w-52 p-2 shadow-sm"
+            >
+              <li>
+                <button onClick={() => setMapKeybindingModalOpen(true)}>
+                  Map keybinding
+                </button>
+              </li>
+            </ul>
+          </div>
+        </div>
+      </div>
+
       {!cabControlState?.Controls?.length && (
         <div className="py-12 text-center">
           <p className="text-base-content/50 text-sm">
@@ -64,15 +96,7 @@ export const CabDebuggerTab = () => {
           </p>
         </div>
       )}
-      {!!cabControlState?.Controls?.length && (
-        <div>
-          <input
-            className="input w-full"
-            placeholder="Search for control(s)"
-            {...register("query")}
-          />
-        </div>
-      )}
+
       <ul className="list bg-base-100 rounded-box shadow-md">
         {sortedControls?.map((controlState) => (
           <CabDebuggerTabControl
@@ -86,6 +110,11 @@ export const CabDebuggerTab = () => {
       <CabDebuggerMapControlModal
         controlState={mapControlModalOpenState?.PropertyName ?? null}
         onClose={handleCloseMapControlModal}
+      />
+
+      <CabDebuggerMapKeybindingModal
+        openState={mapKeybindingModalOpen}
+        onClose={handleCloseMapKeybindingModal}
       />
     </div>
   );