+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## v1.17.0
 - Add support for Running Train
+- Add simple map keybinding form
 
 ## v1.16.1
 - Use HID path matching to optionally find serial